// frontend/src/app/api/generated/model/guest-details.ts
/**
 * Excalibur Gym API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GuestCheckIn } from './guest-check-in';
import { GuestPass } from './guest-pass';
import { GuestRegistrationType } from './guest-registration-type';


export interface GuestDetails { 
    id: string;
    fullName: string;
    registrationType: GuestRegistrationType;
    createdAt: string;
    activeGymPass: boolean;
    checkedInToday: boolean;
    currentPassName?: string | null;
    currentPassValidUntil?: string | null;
    passes: Array<GuestPass>;
    visits: Array<GuestCheckIn>;
}