// frontend/src/app/api/generated/model/guest-pass.ts
/**
 * Excalibur Gym API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PaymentMethod } from './payment-method';


export interface GuestPass { 
    id: string;
    productName: string;
    purchasedAt: string;
    validFrom: string;
    validUntil: string;
    totalEntries?: number | null;
    remainingEntries?: number | null;
    pricePaid: number;
    paymentMethod: PaymentMethod;
    issuedByEmployeeName: string;
    status: GuestPassStatusEnum;
}
export enum GuestPassStatusEnum {
    Active = 'ACTIVE',
    Pending = 'PENDING',
    Expired = 'EXPIRED',
    Invalidated = 'INVALIDATED'
};