// frontend/src/app/api/generated/model/payment-method.ts
/**
 * Excalibur Gym API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum PaymentMethod {

    Cash = 'CASH',

    BankCard = 'BANK_CARD',

    PrepaidBalance = 'PREPAID_BALANCE'
}